remove Lua support
Summary:
For some obscure reason Lua support depends on python2, which is going to be removed.

https://fb.workplace.com/groups/311767668871855/permalink/4219593711422545/

Since the Lua interface is not used in any active code it seems, it's easier to just remove the Lua interface.

It also removes the Faiss Lua dep in the few occurrences where it is used (omry see recog-eval).

Reviewed By: wickedfoo

Differential Revision: D23865458

fbshipit-source-id: 4149517af18acce29179d04152c7364c2548efa0