Disable tensor core usage on V100 GPUs / CUDA 10
Summary:
Tensor core usage on V100 + CUDA 10 for sgemm f32 x f32 = f32 seems to allow demotion of the inputs to f16 (wtf?!), resulting in an unacceptable loss of precision. All accumulation in Faiss is f32, and the use cases for f16 x f16 = f32 are opt-in and relatively small in practice I believe.

For A100 or CUDA 11, there is an option to only allow tensor core usage if we guarantee preservation of precision, which we instead prefer.

Reviewed By: beauby

Differential Revision: D24348944

fbshipit-source-id: d22cfaa233d21ee9c20974914ad155dab8c901fd