Suppress -Wpedantic (#1888)
Summary:
Current `faiss` contains some codes which will be warned by compilers when we will add some compile options like `-Wall -Wextra` .
IMHO, warning codes in `.cpp` and `.cu` doesn't need to be fixed if the policy of this project allows warning.
However, I think that it is better to fix the codes in `.h` and `.cuh` , which are also referenced by `faiss` users.

Currently it makes a error to `#include` some faiss headers like `#include<faiss/IndexHNSW.h>` when we compile the codes with `-pedantic-errors` .
This PR fix this problem.
In this PR, for the reasons above, we fixed `-Wpedantic` codes only in `.h` .

This PR doesn't change `faiss` behavior.

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/1888

Reviewed By: wickedfoo

Differential Revision: D28506963

Pulled By: beauby

fbshipit-source-id: cbdf0506a95890c9c1b829cb89ee60e69cf94a79