Faiss GPU: remove d2h copy from PQ MM debugging
Summary:
If the number of dimensions per sub-quantizer is not in the specialized list, it falls back to the generalized batch GEMM implementation.

When I implemented this, I had in a d2h copy so I could look at the computed distances. I removed the debugging code but not this copy.

Prior to this, PQ16 on 1024 dims was 6x slower than PQ32. Now, it is only 1.5x slower (it is slower because there is a higher number of dims per sub-q, despite there being more sub-qs).

Reviewed By: beauby

Differential Revision: D34526043

fbshipit-source-id: de6f70f0f0b91608eb6ae2a05da2af812546e4bc