Automatic type conversions for Python API (#2274)
Summary:
Pull Request resolved: https://github.com/facebookresearch/faiss/pull/2274

All input matrices needed to be of the correct type and to be C-contiguous. This diff passes the main entry points of the api through `np.ascontiguousarray` so that the function parameters are transparently converted to the suitable format if needed.

We did not have this before because users need to be made aware of the performance impact, but it seems that maybe usability is more useful.

This diff is an alternative to

D35007365
https://github.com/facebookresearch/faiss/pull/2250

Reviewed By: beauby

Differential Revision: D35009612

fbshipit-source-id: fa0d5cfdfbff6b0916d47bd33c620e3ca9d5dd40