Faiss GPU IVF large query batch fix (#2572)
Summary:
Pull Request resolved: https://github.com/facebookresearch/faiss/pull/2572

This is a fix to https://github.com/facebookresearch/faiss/issues/2561

namely, GpuIndexIVFFlat didn't work for query batch sizes larger than 65536, due to that being the maximum grid Y dimension.

This code fixes the IVFFlat code to instead perform a grid loop over the queries as needed for >65536, and in other cases which are not currently violated by large batch queries but we're simply passing in some array size as a grid Y parameter, I added asserts to catch this for the future.

Added two tests for IVFPQ and IVFFlat for the large batch case. The IVFPQ large batch test passed before, but IVFFlat reproduced the same issue as seen in GH issue 2561.

Reviewed By: alexanderguzhva

Differential Revision: D41184878

fbshipit-source-id: 92a87bd095319d6dcd73c76ba9044c019b8ca71c