remove unnecessary nbits_per_idx check
Summary:
`IndexIVFPQ` doesn't seem to depend on single-byte code words, and
encoding / decoding is abstracted away out of it.

There is some reachable logic that supports only `pq.nbits == 8`
(polysemous training is the only one I found), but those are themselves
are gated, and missing support is an implementation detail (also, the
existing `<=` is wrong even if the intention was to gate on polysemous
training support).

Reviewed By: luciang

Differential Revision: D41483414

fbshipit-source-id: 06f471c25293e01242d7bab37ff54e709edc710b