Relax IVF AQ FastScan (#2940)
Summary:
Pull Request resolved: https://github.com/facebookresearch/faiss/pull/2940

This test fails on some occasions.
After investigation it turns out this is due to non reproducible behavior IndexIVFFastScan::search_implem_14 with a parallel loop, where there are ties in the resutls (ie. the resulting distances are the same but not the ids).
As a workaround I relaxed the test slightly.
+ a fix in the checksum function.

Reviewed By: algoriddle

Differential Revision: D47229086

fbshipit-source-id: 55e53bcfe47cf33041cc7fd5691b5de65067ce0f