Using expanded distance forms in `RaftFlatIndex.cu` (#3021)
Summary:
This is a minor bug that comes with a perf impact. The classic FAISS `FlatIndex` always uses expanded form of distance computation even though an argument `exactDistances` is provided. `RaftFlatIndex` was using this argument to determine whether the computation should be exhaustive.

This PR includes one additional change to eagerly initialize the `cublas_handle` on the `device_resources` instance when it's created.

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/3021

Reviewed By: pemazare

Differential Revision: D48739660

Pulled By: mdouze

fbshipit-source-id: a361334eb243df86c169c69d24bb10fed8876ee9