Add conda bin to path early in the cmake GitHub action (#3512)
Summary:
Pull Request resolved: https://github.com/facebookresearch/faiss/pull/3512

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/3510

GitHub hosted runners some with the build-essentials package pre-installed, self-hosted runners on AWS do not have this package. This made it all steps other than the `all targets` one fall back to the system executables which unintentially worked on GitHub hosted runners but not on the self-hosted ones. This diff fixes it by pulling the line that adds conda bin to path early in the cmake build action.

Reviewed By: asadoughi

Differential Revision: D58513853

fbshipit-source-id: 23e95459e0031c96bd142515db07d1b700d713cf