Move static functions to header file (#3757)
Summary:
Pull Request resolved: https://github.com/facebookresearch/faiss/pull/3757

In the telemetry wrapper, we need to wrap read_index to return wrapped index structs. D61049751

This read_index wrapper calls several static functions. These are not callable outside a C++ file. Thus this diff changes them to non static and declares them in the header file. Then the wrapper is able to call them.

Reviewed By: asadoughi

Differential Revision: D61282004

fbshipit-source-id: 2c8b2ded169577aa6eecdf1edc7483b0ef5f0665