Add index binary to telemetry (#4001)
Summary:
Pull Request resolved: https://github.com/facebookresearch/faiss/pull/4001

It was mentioned in S461104 chat to cover all index types. This adds Binary for telemetry as well as the reverse factory string for binary indexes, which we did not support before.

Unit test covers 4 ways of reading a binary index.

The reverse factory string util is not complete. The remaining binary index types could get added later.

Reviewed By: asadoughi

Differential Revision: D65102643

fbshipit-source-id: 52f1053bda59e427a081369ada80265b67e55bd4